New file for inbox elements
