+++ Canvas/CanvasUITests/Inbox/InboxTests.swift
@@ -17,58 +17,6 @@
 import XCTest
 import TestsFoundation
 
-enum Inbox {
-    static var filterButton: Element {
-        return app.find(id: "inbox.filterByCourse")
-    }
-
-    static var newMessageButton: Element {
-        return app.find(id: "inbox.new-message")
-    }
-
-    static func filterOption(_ text: String) -> Element {
-        return app.find(labelContaining: "Assignments")
-    }
-
-    static func message(id: String) -> Element {
-        return app.find(id: "inbox.conversation-\(id)")
-    }
-}
-
-enum NewMessage {
-    static var selectCourseButton: Element {
-        return app.find(id: "compose.course-select")
-    }
-
-    static var addRecipientButton: Element {
-        return app.find(id: "compose.add-recipient")
-    }
-}
-
-enum MessageCourseSelection {
-    static func course(id: String) -> Element {
-        return app.find(id: "inbox.course-select.course-\(id)")
-    }
-}
-
-enum MessageRecipientsSelection {
-    static func messageAllInCourse(courseID: String) -> Element {
-        return app.find(id: "branch_course_\(courseID)")
-    }
-
-    static func studentsInCourse(courseID: String) -> Element {
-        return app.find(id: "course_\(courseID)_students")
-    }
-
-    static func messageAllStudents(courseID: String) -> Element {
-        return app.find(id: "branch_course_\(courseID)_students")
-    }
-
-    static func student(studentID: String) -> Element {
-        return app.find(id: studentID)
-    }
-}
-
 class InboxTests: CanvasUITests {
     override var user: User? { return .student1 }
